requirements.txt
python-kraken-sdk
websocket-client
kafka-python